# api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flashcards Open Source App External AI-Agent API",
    "version": "v1",
    "description": "Public machine-facing contract for external AI agents. This document intentionally excludes first-party/internal routes such as sync transport, local chat runtime endpoints, and human-only agent connection management even when those routes remain live for web or iOS clients."
  },
  "servers": [
    {
      "url": "https://api.flashcards-open-source-app.com/v1",
      "description": "Canonical API host for discovery, bootstrap, docs, and external agent tools"
    },
    {
      "url": "https://auth.flashcards-open-source-app.com",
      "description": "Auth host for email OTP bootstrap"
    }
  ],
...
  "paths": {
...
    "/api/agent/send-code": {
      "post": {
        "tags": ["auth"],
        "summary": "Send OTP code",
        "description": "Starts the external AI-agent auth flow by emailing an 8-digit OTP code and returning a short opaque otpSessionToken.",
        "servers": [
          {
            "url": "https://auth.flashcards-open-source-app.com",
            "description": "Auth host for email OTP bootstrap"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSendCodeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP challenge created"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Use the header format: Authorization: ApiKey $FLASHCARDS_OPEN_SOURCE_API_KEY"
      }
    },
    "schemas": {
...
      "AgentSendCodeInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      },
...
    }
  }
}